model/requirements.txt
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.2.0
matplotlib>=3.6.0
seaborn>=0.12.0
jupyter>=1.0.0
requests>=2.28.0
kaggle>=1.5.12
python-dotenv>=0.21.0
xgboost>=1.7.0
lightgbm>=3.3.0
joblib>=1.2.0
pyyaml>=6.0
pytest>=7.0.0
pytest-cov>=4.0.0